requirements.txt
psutil
pyyaml
PyQt5